Integrar Eureka y refactorizar los tipos de ID de usuario
Se agregó la configuración del cliente de Eureka y sus dependencias a todos los servicios, e introdujo un nuevo servicio de descubrimiento con la configuración del servidor de Eureka. Se refactorizaron los campos de ID de usuario y los genéricos del repositorio de int a long en los modelos y servicios relacionados con el usuario para mayor consistencia. Se eliminaron los archivos de prueba redundantes y se actualizaron las propiedades de la aplicación para el registro del servicio.

+++ auth-service/src/main/java/com/example/loopie/Users/service/UserService.java
@@ -26,7 +26,7 @@ public class UserService implements UserDetailsService {
         return repository.findAll();
     }
 
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Usuario no encontrado"));
     }
@@ -36,7 +36,7 @@ public class UserService implements UserDetailsService {
         return repository.save(user);
     }
 
-    public User updateUser(int id, User user) {
+    public User updateUser(Long id, User user) {
         User dbUser = getUserById(id);
         dbUser.setUsername(user.getUsername());
         dbUser.setEmail(user.getEmail());
@@ -45,7 +45,7 @@ public class UserService implements UserDetailsService {
         return repository.save(dbUser);
     }
 
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         repository.deleteById(id);
     }
